A couple component and file renames
